Добавлены параметры лидерборда

+++ Samples/Data/LeaderboardEntityElement.cs
@@ -9,7 +9,7 @@ public class LeaderboardEntityElement : MonoBehaviour
     [SerializeField] private TMP_Text _name;
     [SerializeField] private TMP_Text _score;
 
-    public void Initialize(string name, int rank, int score)
+    public void Initialize(string name, long rank, long score)
     {
         _name.text = name;
         _rank.text = rank.ToString();
